back_end/pyproject.toml
[tool.poetry]
name = "clinical_trials_backend"
version = "1.1.0"
description = ""
authors = ["abdullah <abdullahwaqar@pm.me>"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
clinicaltrials = { path = "./clinical_trials_core", develop = true }
fastapi = "^0.115.4"
hypercorn = "*"
loguru = "*"
orjson = "*"
python = "^3.11"
python-dotenv = "*"
uvloop = "^0.21.0"
sqlmodel = "^0.0.22"
psycopg2-binary = "^2.9.7"
pyjwt = "^2.8.0"
bcrypt = "^4.0.1"
aiosmtplib = "^2.0.2"
python-multipart = "^0.0.6"
boto3 = "^1.34.14"
boto3-stubs = { extras = ["essential"], version = "^1.34.21" }
email-validator = "^2.1.0.post1"
pydantic = "^2.9.2"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
redis = "^5.0.1"
celery = { extras = ["redis"], version = "^5.4.0" }
azure-storage-blob = "^12.23.1"
httpx = "0.28.0"
phonenumbers = "^8.13.49"
pdfkit = "^1.0.0"
wordcloud = "^1.9.4"
pillow = "^11.0.0"
nltk = "^3.8.1"
jinja2 = "^3.1.5"
plotly = "^5.24.1"
kaleido = "0.2.1"

[tool.poetry.group.test.dependencies]
asgi-lifespan = "^2.1.0"
alt-pytest-asyncio = "^0.9.1"
httpx = "^0.28.0"
pip-audit = "^2.7.3"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
tox = "^4.23.2"

[tool.poetry.group.dev.dependencies]
blue = "*"
mypy = "*"
alembic =  "^1.14.0"
[tool.pyright]

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff]
line-length = 180

select = [
    "ARG",
    "B",
    "C4",
    "E",
    "ERA",
    "F",
    "I",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "RET",
    "S",
    "T10",
    "UP",
    "W",
]
ignore = ["B008"]
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["ARG", "PLR2004", "S"]
"smoke_test.py" = ["S"]

[tool.ruff.pylint]
max-args = 10
max-statements = 60
